+++ martin/src/config/file/file_config.rs
@@ -5,6 +5,7 @@ use std::path::{Path, PathBuf};
 
 use log::{info, warn};
 use martin_core::cache::OptMainCache;
+use martin_core::config::IdResolver;
 use martin_core::config::OptOneMany::{self, Many, One};
 use martin_core::tiles::BoxedSource;
 use serde::{Deserialize, Serialize};
@@ -13,7 +14,6 @@ use url::Url;
 use crate::config::file::ConfigFileError::{
     InvalidFilePath, InvalidSourceFilePath, InvalidSourceUrl, IoError,
 };
-use crate::utils::IdResolver;
 use crate::{MartinError, MartinResult};
 
 pub type ConfigFileResult<T> = Result<T, ConfigFileError>;
