+++ martin/src/config/file/postgres/builder.rs
@@ -4,6 +4,7 @@ use std::collections::HashSet;
 use futures::future::join_all;
 use itertools::Itertools as _;
 use log::{debug, error, info, warn};
+use martin_core::config::IdResolver;
 use martin_core::config::OptBoolObj::{Bool, NoValue, Object};
 use martin_core::config::OptOneMany::NoVals;
 use martin_core::tiles::BoxedSource;
@@ -21,7 +22,6 @@ use crate::config::file::postgres::{
     PostgresConfig, PostgresInfo, TableInfo, TableInfoSources,
 };
 use crate::config::file::{ConfigFileError, ConfigFileResult};
-use crate::utils::IdResolver;
 
 /// Builder for [`PostgresSource`]' auto-discovery of functions and tables.
 #[derive(Debug)]
