--- martin/src/utils/mod.rs
@@ -1,8 +1,0 @@
-mod error;
-pub use error::*;
-
-mod id_resolver;
-pub use id_resolver::IdResolver;
-
-mod utilities;
-pub use utilities::*;
